refactor(mls): extract MlsService trait from OpenMlsService impl
