+++ src/core/api/mod.rs
@@ -1,7 +1,7 @@
 //! Core API for MLS group operations.
 //!
 //! All functions operate on a [`crate::core::Group`] (app-level state) plus
-//! [`crate::mls_crypto::MlsService`] (MLS crypto). Lifecycle, message
+//! [`crate::mls_crypto::OpenMlsService`] (MLS crypto). Lifecycle, message
 //! building, inbound routing, steward candidate creation, and freeze
 //! finalization are grouped by submodule.
